Change main unit to always be the 0 (and more) (#870)
