+++ renovate.json
@@ -21,6 +21,14 @@
       "datasourceTemplate": "github-releases",
       "depNameTemplate": "element-hq/synapse",
       "versioningTemplate": "semver-coerced"
+    },
+    {
+      "fileMatch": ["(^|/)src/charm\\.py$"],
+      "description": "Update Synapse version in src/charm.py",
+      "matchStrings": ["SYNAPSE_VERSION\\s*=\\s*\"(?<currentValue>.*?)\""],
+      "datasourceTemplate": "github-releases",
+      "depNameTemplate": "element-hq/synapse",
+      "versioningTemplate": "semver-coerced"
     }
 
   ],
