+++ src/pebble.py
@@ -51,7 +51,6 @@ def check_synapse_alive(charm_state: CharmState) -> ops.pebble.CheckDict:
     """
     check = Check(synapse.CHECK_ALIVE_NAME)
     check.override = "replace"
-    check.level = "alive"
     check.tcp = {"port": synapse.SYNAPSE_PORT}
     experimental_alive_check = charm_state.synapse_config.experimental_alive_check
     if experimental_alive_check:
@@ -70,7 +69,6 @@ def check_synapse_ready() -> ops.pebble.CheckDict:
     """
     check = Check(synapse.CHECK_READY_NAME)
     check.override = "replace"
-    check.level = "ready"
     check.timeout = "20s"
     check.period = "2m"
     check.threshold = 5
@@ -108,7 +106,6 @@ def check_nginx_ready() -> ops.pebble.CheckDict:
     """
     check = Check(synapse.CHECK_NGINX_READY_NAME)
     check.override = "replace"
-    check.level = "ready"
     check.http = {"url": f"http://localhost:{synapse.SYNAPSE_NGINX_PORT}/health"}
     return check.to_dict()
 
@@ -121,7 +118,6 @@ def check_mjolnir_ready() -> ops.pebble.CheckDict:
     """
     check = Check(synapse.CHECK_MJOLNIR_READY_NAME)
     check.override = "replace"
-    check.level = "ready"
     check.http = {"url": f"http://localhost:{synapse.MJOLNIR_HEALTH_PORT}/healthz"}
     check.timeout = "10s"
     check.threshold = 5
@@ -432,7 +428,7 @@ def _pebble_layer(charm_state: CharmState, is_main: bool = True) -> ops.pebble.L
                 "startup": "enabled",
                 "command": command,
                 "environment": synapse.get_environment(charm_state),
-            }
+            },
         },
         "checks": {
             synapse.CHECK_ALIVE_NAME: check_synapse_alive(charm_state),
@@ -440,23 +436,6 @@ def _pebble_layer(charm_state: CharmState, is_main: bool = True) -> ops.pebble.L
         },
     }
     return typing.cast(ops.pebble.LayerDict, layer)
-
-
-def _pebble_layer_without_restart(charm_state: CharmState) -> ops.pebble.LayerDict:
-    """Return a dictionary representing a Pebble layer without restart.
-
-    Args:
-        charm_state: Instance of CharmState
-
-    Returns:
-        pebble layer
-    """
-    new_layer = _pebble_layer(charm_state)
-    new_layer["services"][synapse.SYNAPSE_SERVICE_NAME]["on-success"] = "ignore"
-    new_layer["services"][synapse.SYNAPSE_SERVICE_NAME]["on-failure"] = "ignore"
-    ignore = {synapse.CHECK_READY_NAME: "ignore"}
-    new_layer["services"][synapse.SYNAPSE_SERVICE_NAME]["on-check-failure"] = ignore
-    return new_layer
 
 
 def _nginx_pebble_layer() -> ops.pebble.LayerDict:
