+++ tests/unit_harness/test_anonymize_user_action.py
