+++ tests/unit_harness/test_charm_state.py
@@ -40,13 +40,16 @@ class SimpleCharm(CharmBaseWithState):
             redis_config=None,
             instance_map_config=None,
             registration_secrets=None,
+            redis_required=None,
+            mjolnir_enabled=None,
         )
 
-    def reconcile(self, charm_state) -> None:
+    def reconcile(self, charm_state, maintenance_status: str = "Configuring Synapse") -> None:
         """Reconcile configuration.
 
         Args:
             charm_state: The charm state.
+            maintenance_status: msg.
         """
