+++ tests/unit_harness/test_matrix_auth_integration.py
