+++ tests/unit_harness/test_synapse_api.py
@@ -232,6 +232,8 @@ def test_override_rate_limit_success(monkeypatch: pytest.MonkeyPatch):
         redis_config=None,
         instance_map_config=None,
         registration_secrets=None,
+        redis_required=None,
+        mjolnir_enabled=None,
     )
     expected_url = (
         f"http://localhost:8008/_synapse/admin/v1/users/@any-user:{server}/override_ratelimit"
@@ -271,6 +273,8 @@ def test_override_rate_limit_error(monkeypatch: pytest.MonkeyPatch):
         redis_config=None,
         instance_map_config=None,
         registration_secrets=None,
+        redis_required=None,
+        mjolnir_enabled=None,
     )
     expected_error_msg = "Failed to connect"
     do_request_mock = mock.MagicMock(side_effect=synapse.APIError(expected_error_msg))
