+++ tests/unit_harness/test_synapse_workload.py
@@ -258,6 +258,8 @@ def test_enable_trusted_key_servers_no_action(config_content: dict[str, typing.A
             synapse_config=synapse_config,
             instance_map_config=None,
             registration_secrets=None,
+            redis_required=None,
+            mjolnir_enabled=None,
         ),
     )
 
@@ -543,6 +545,8 @@ def test_enable_smtp_success(config_content: dict[str, typing.Any]):
         instance_map_config=None,
         synapse_config=synapse_config,
         registration_secrets=None,
+        redis_required=None,
+        mjolnir_enabled=None,
     )
 
     synapse.enable_smtp(config_content, charm_state)
@@ -720,6 +724,8 @@ def test_block_non_admin_invites(config_content: dict[str, typing.Any]):
         media_config=None,
         instance_map_config=None,
         registration_secrets=None,
+        redis_required=None,
+        mjolnir_enabled=None,
     )
 
     synapse.block_non_admin_invites(config_content, charm_state)
@@ -756,6 +762,8 @@ def test_publish_rooms_allowlist_success(config_content: dict[str, typing.Any]):
         media_config=None,
         instance_map_config=None,
         registration_secrets=None,
+        redis_required=None,
+        mjolnir_enabled=None,
     )
 
     synapse.enable_room_list_publication_rules(config_content, charm_state)
@@ -870,6 +878,8 @@ def test_invite_checker_policy_rooms(config_content: dict[str, typing.Any]):
         media_config=None,
         instance_map_config=None,
         registration_secrets=None,
+        redis_required=None,
+        mjolnir_enabled=None,
     )
 
     synapse.enable_synapse_invite_checker(config_content, charm_state)
@@ -913,6 +923,8 @@ def test_invite_checker_blocklist_allowlist_url(config_content: dict[str, typing
         media_config=None,
         instance_map_config=None,
         registration_secrets=None,
+        redis_required=None,
+        mjolnir_enabled=None,
     )
 
     synapse.enable_synapse_invite_checker(config_content, charm_state)
