+++ tox.ini
@@ -91,6 +91,7 @@ deps =
     cosl
     coverage[toml]
     pytest
+    ops[testing]
     -r{toxinidir}/requirements.txt
 commands =
     coverage run --source={[vars]src_path} \
